+++ build.gradle
@@ -1,26 +1,36 @@
 buildscript {
     ext.versions = [
+            // Virgil
             virgilSdk : '5.0.4',
-            kotlin    : '1.3.0-rc-190',
+            pythia : '0.2.0',
+            keyknox : '0.1.0',
+
+            // Kotlin
+            kotlin    : '1.3.0',
             coroutines: '1.0.0-RC1',
-            klaxon    : '3.0.5',
-            junit     : '5.2.0',
-            dokka     : '0.9.17',
+
+            // Gradle
+            gardle    : '3.2.1',
+
+            // Android
             android   : '4.1.1.4',
+
+            // Docs
+            dokka     : '0.9.17',
+
+            // Tests
+            junit     : '5.2.0',
     ]
     repositories {
         google()
         jcenter()
         mavenLocal()
         mavenCentral()
-        maven {
-            url 'https://dl.bintray.com/kotlin/kotlin-eap'
-        }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.1'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:${versions.kotlin}"
-        classpath "org.jetbrains.dokka:dokka-gradle-plugin:${versions.dokka}"
+        classpath "com.android.tools.build:gradle:$versions.gardle"
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$versions.kotlin"
+        classpath "org.jetbrains.dokka:dokka-gradle-plugin:$versions.dokka"
     }
 }
