+++ common/build.gradle
@@ -12,9 +12,6 @@ group 'com.virgilsecurity'
 version '0.1.1-rc'
 
 dependencies {
-    // Inner dependencies
-    implementation project(':common')
-
     // Virgil
     compileOnly "com.virgilsecurity.sdk:crypto:$versions.virgilSdk"
     implementation "com.virgilsecurity.sdk:crypto-android:$versions.virgilSdk@aar"
