+++ ethreeCoroutines/build.gradle
@@ -9,7 +9,7 @@ apply plugin: 'maven'
 apply plugin: 'org.jetbrains.dokka'
 
 group 'com.virgilsecurity'
-version '0.1.1-rc'
+version '0.1.0-alpha'
 
 dependencies {
     // Inner dependencies
