+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':tests', ':ethree'
+include ':tests', ':ethree', ':ethreeCoroutines', ':common', ':testscoroutines'
